+++ tests/unit/conftest.py
@@ -20,6 +20,7 @@ from graphdatascience.query_runner.query_mode import QueryMode
 from graphdatascience.server_version.server_version import ServerVersion
 from graphdatascience.session.aura_graph_data_science import AuraGraphDataScience
 from graphdatascience.session.dbms_connection_info import DbmsConnectionInfo
+from graphdatascience.session.session_lifecycle_manager import SessionLifecycleManager
 
 # Should mirror the latest GDS server version under development.
 DEFAULT_SERVER_VERSION = ServerVersion(2, 10, 0)
@@ -229,7 +230,7 @@ def aura_gds(runner: CollectingQueryRunner, mocker: MockerFixture) -> Generator[
         session_bolt_connection_info=DbmsConnectionInfo("address", "some", "auth"),
         arrow_authentication=UsernamePasswordAuthentication("some", "auth"),
         db_endpoint=DbmsConnectionInfo("address", "some", "auth"),
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
     )
     yield aura_gds