+++ tests/unit/session/test_aura_graph_data_science.py
@@ -1,8 +1,11 @@
 from pytest_mock import MockerFixture
 
 from graphdatascience import ServerVersion
+from graphdatascience.query_runner.neo4j_query_runner import Neo4jQueryRunner
 from graphdatascience.query_runner.query_mode import QueryMode
 from graphdatascience.session.aura_graph_data_science import AuraGraphDataScience
+from graphdatascience.session.session_lifecycle_manager import SessionLifecycleManager
+from graphdatascience.session.session_v2_endpoints import SessionV2Endpoints
 from tests.unit.conftest import CollectingQueryRunner
 
 
@@ -11,7 +14,7 @@ def test_remote_projection_configuration(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -48,7 +51,7 @@ def test_remote_projection_defaults(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -77,7 +80,7 @@ def test_remote_algo_write(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -99,7 +102,7 @@ def test_remote_algo_write_configuration(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -126,7 +129,7 @@ def test_remote_graph_write(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -151,7 +154,7 @@ def test_remote_graph_write_configuration(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -179,7 +182,7 @@ def test_run_cypher_write(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -197,7 +200,7 @@ def test_run_cypher_read(mocker: MockerFixture) -> None:
     query_runner = CollectingQueryRunner(v)
     gds = AuraGraphDataScience(
         query_runner=query_runner,
-        delete_fn=lambda: True,
+        session_lifecycle_manager=mocker.Mock(spec=SessionLifecycleManager),
         gds_version=v,
         v2_endpoints=mocker.Mock(),
         authenticated_arrow_client=mocker.Mock(),
@@ -213,3 +216,40 @@ def test_run_cypher_read(mocker: MockerFixture) -> None:
         "mode": QueryMode.READ,
         "retryable": False,
     }
+
+
+def test_verify_connectivity(mocker: MockerFixture) -> None:
+    v = ServerVersion(9, 9, 9)
+    query_runner = mocker.Mock(spec=Neo4jQueryRunner)
+    session_lifecycle_manager = mocker.Mock(spec=SessionLifecycleManager)
+    v2_endpoints = mocker.Mock(spec=SessionV2Endpoints)
+    gds = AuraGraphDataScience(
+        query_runner=query_runner,
+        session_lifecycle_manager=session_lifecycle_manager,
+        gds_version=v,
+        v2_endpoints=v2_endpoints,
+        authenticated_arrow_client=mocker.Mock(),
+    )
+
+    gds.verify_connectivity()
+
+    session_lifecycle_manager.verify_health.assert_called_once()
+    v2_endpoints.verify_session_connectivity.assert_called_once()
+    v2_endpoints.verify_db_connectivity.assert_called_once()
+
+
+def test_delete(mocker: MockerFixture) -> None:
+    v = ServerVersion(9, 9, 9)
+    query_runner = CollectingQueryRunner(v)
+    session_lifecycle_manager = mocker.Mock(spec=SessionLifecycleManager)
+    gds = AuraGraphDataScience(
+        query_runner=query_runner,
+        session_lifecycle_manager=session_lifecycle_manager,
+        gds_version=v,
+        v2_endpoints=mocker.Mock(),
+        authenticated_arrow_client=mocker.Mock(),
+    )
+
+    gds.delete()
+
+    session_lifecycle_manager.delete.assert_called_once()